feat: Improve codebase
