+++ src/components/big-label/index.js
@@ -1,0 +1,1 @@
+export { default } from './BigLabel';
