+++ src/pages/yokai-card/index.js
@@ -1,0 +1,1 @@
+export { default } from './YokaiCard';
